BIRTHDAY CHOCOLATE
Lily has a chocolate bar that she wans to share with Ron for his birthday.
Each of the squares has an integer on it. She decides to share a contiguous
segment of the bar selected such that the length of the segment matches Rom's
birth month and the sum of the integers on the square is equal to his birth
day. You must determine how many ways she can divide the chocolate.