Angry Professor
A discrete Mathematics Professor has a class of Students. Frustrated with their
lack of discipline, he decides to cancel class if fewer than some number of
students are present when class starts. Arrival times go from on time
(arrivalTime <= 0) to arrived time late (arrivalTime > 0).

Given the arrival time of each student and a threshold number of attendees,
determine if the class is canceled.