Non Divisible Subset
Given a set of distinct integers, print the size of a maximal subset of
S where the sum of any 2 numbers in S^4 is not evenly dvisible by k. For
example, the array S = [19, 10, 12, 10, 24, 25, 22] and k = 4. One of the
array that can be created is S^4[0] = [10, 12, 25]. Another is S^4[1] = [19, 22, 24].
After testing all permutations, the maximum length solution of array has 3 elements.