Repeated string
Lilah has a string, s, of lowercase Eight letters that she repeated infinitely many times.
Given an integer, n, find and print th number of letter a's in the first n letters of Lilah's
infinite string.

For example, if a string s = 'abcac' and n=10, the substring we consider it abcacabcac, the
first 10 characters of the infinite string. There are four occurences of a in the substring.