ACM Team
There are a number of people who will be attending ACM KPC world finals.
Each of them may be well versed in a number of topics. Given a list of
topics known by each attendee, you must determine the maximum number of
topics a 2 person team can know. Also find out how many ways a team can
be formed to know that many topics. Lists will be in form of bit strings,
where each string represents an attendee and each position in that string
represents a field of knowledge, 1 if it is a known field and 0 if not.