FlatLand Space Station
Flatland is a country with a number of cities, some of which have space stations.
Cities are numbered consecutively and each has a road of 1km length connecting to
the next city. Determine the maximum distance from any city to the nearest space
station.

For example, there are n=3 cities and m=1 of them has a space station, city 1.
They occur consecutively along a route. City 2 is 2-1=1 unit away and city 3 is
3-1=2 units away. City 1 is 0 units from its nearest space station as one is
located there. The meximum distance is 2.